requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# MLflow
mlflow==2.8.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-multipart==0.0.6
requests==2.31.0
seaborn==0.13.2
scipy==1.11.4
matplotlib==3.8.2

# Application Insights
opencensus-ext-azure==1.1.9
opencensus-ext-requests==0.12.1
opencensus-ext-fastapi==0.6.0  # préciser une version stable
